update price field

+++ store/models.py
@@ -15,7 +15,7 @@ class Categories(models.Model):
 class Products(models.Model):
     title = models.CharField(max_length=255, verbose_name='Название товара')
     description = models.TextField(verbose_name='Описание товара')
-    price = models.IntegerField(verbose_name='Цена')
+    price = models.DecimalField(max_digits=8, decimal_places=2, verbose_name='Цена') 
     category = models.ForeignKey(Categories, on_delete=models.CASCADE, related_name='category', verbose_name='Категория')
     create_at = models.DateTimeField(auto_now_add=True, verbose_name='Дата добавления товара')
     update_at = models.DateTimeField(auto_now=True, verbose_name='Дата обновления товара')
